refactor/fix: `unload_models_from_vram` more often
With some recent comfyui changes it appears that the logic prior to this commit was not aggressive enough to avoid OOMs with relying on comfyui's internal decision making alone.

This commit causes the worker to unload models from VRAM immediately after an inference result (if it is not about to be used) and right before post processing.

Post processing as implemented today almost always overestimates the amount of free VRAM, and tends to cause OOMs or shared memory usage (on window) so more proactively unloading the model should help minimize that problem.